Posts have been removed in Jekyll 3.x.
In 3.x, posts have become documents. But we still want to handle posts
in Jekyll 2.x and earlier, since documents at those times don't have a
ton of information (e.g. no date).

Additionally in Jekyll 3.x, you must loop on `site.posts.docs` to avoid
a deprecation error, so we use that if it's available.